Conflicts:
	source/commands/makecontigscommand.h

+++ source/calculators/accuracy.hpp
@@ -17,7 +17,7 @@ class Accuracy : public ClusterMetric  {
     
 public:
     Accuracy() : ClusterMetric("accuracy") {};
-    double getValue( long long tp,  long long tn,  long long fp,  long long fn);
+    double getValue(double tp,  double tn,  double fp,  double fn);
     string getCitation() { return "http://www.mothur.org/wiki/Accuracy"; }
     
 private:
