+++ source/calculators/tp.hpp
@@ -17,7 +17,7 @@ class TP : public ClusterMetric  {
     
 public:
     TP() : ClusterMetric("tp") {};
-    double getValue( long long tp,  long long tn,  long long fp,  long long fn);
+    double getValue(double tp,  double tn,  double fp,  double fn);
     string getCitation() { return "http://www.mothur.org/wiki/TP"; }
     
 private:
